Struct radeco_lib::analysis::dom::domtree::DomInfo

pub struct DomInfo {
    // some fields omitted
}

Wrapper struct that consolidates the Dom and PostDom information for a Graph.

Methods

impl DomInfo

fn new() -> DomInfo

fn build_dom_tree<N, E>(&mut self, g: &Graph<N, E>, i: NodeIndex) where N: Clone, E: Clone

Constructs the DomTree and computes the dominance frontier for g.

fn build_postdom_tree<N, E>(&mut self, g: &Graph<N, E>, i: NodeIndex) where N: Clone, E: Clone

Constructs the Post-DomTree and computes the postdominance frontier for g.

fn doms(&self, i: NodeIndex) -> Vec<NodeIndex>

Returns all the dominators of the node with NodeIndex i.

fn idom(&self, i: NodeIndex) -> NodeIndex

Returns the immediate dominator of the node with NodeIndex i.

fn dom_frontier(&self, i: NodeIndex) -> HashSet<NodeIndex>

Returns all nodes in the dominance frontier of i.

fn postdoms(&self, i: NodeIndex) -> Vec<NodeIndex>

Returns all postdominators of i.

fn ipostdom(&self, i: NodeIndex) -> NodeIndex

Returns the immediate postdominator of i.

fn postdom_frontier(&self, i: NodeIndex) -> HashSet<NodeIndex>

Returns all nodes in the postdominance frontier of i.