Module radeco_lib::analysis::valueset

This module offers traits for valuesets and implementations for valuesets of 64 bit integers using different compact representations, including methods for intersection and union. (Currently unused)

Structs

KnownBits
Value set of u64 integers with certain bits set/cleared.

SIntRange
Value set of signed values between min and max.

UIntMultiple
Value set of unsigned values that satisfy value % modulus = residue.

UIntRange
Value set of unsigned values between min and max.

Traits

ScannableSet
A value set over a totally ordered domain with methods to query the next greater or smaller value

ValueSet
A value set. Some implementations introduce false positives when intersected or unified with each other.