Trait radeco_lib::analysis::valueset::ScannableSet

pub trait ScannableSet<T: Ord>: ValueSet<T> {
    fn scan_up(&self, value: T) -> Option<T>;
    fn scan_dn(&self, value: T) -> Option<T>;
}

A value set over a totally ordered domain with methods to query the next greater or smaller value

Required Methods

fn scan_up(&self, value: T) -> Option<T>

Return Some(x) where x ≥ value and x ∈ self

fn scan_dn(&self, value: T) -> Option<T>

Return Some(x) where x ≤ value and x ∈ self