Struct radeco_lib::frontend::parser::Parser

pub struct Parser<'a> {
    // some fields omitted
}

Methods

impl<'a> Parser<'a>

fn new(config: Option<ParserConfig<'a>>) -> Parser<'a>

fn set_register_profile(&mut self, reg_info: &LRegInfo)

fn set_flags(&mut self, flags: &Vec<LFlagInfo>)

fn new_mreg_info(&self, info: LRegProfile) -> MRegInfo

fn parse_opinfo(&mut self, opinfo: &LOpInfo) -> Result<(), ParseError>

fn parse_str(&mut self, esil: &str) -> Result<(), ParseError>

fn emit_insts(&mut self) -> Vec<MInst>

Emit-instructions converts certain compound instructions to simpler instructions. For example, the sequence of instructions, none if zf (OpIf) jump addr (OpJmp) is converted to a single conditional jump as, none if zf jump addr (OpCJmp)