Module radeco_lib::middle

Components processing data in SSA form.

Modules

cfg
This module is used to convert the instructions generated from esil into a CFG.

dce
Dead code elimination

display
Module to implement the display traits for the IR.

dot
Graph visualization traits and functions to emit dot code.

ir
Enumerations and structs to represent instructions. Shared by the SSA representation and the intermediate representation of the ESIL parser.

phiplacement
Implements the SSA construction algorithm described in "Simple and Efficient Construction of Static Single Assignment Form"

regfile
Contains the struct SubRegisterFile which extends PhiPlacers functionality by reads and writes to partial registers.

ssa
Holds traits and structs for the ssa form