Trait radeco_lib::middle::dot::GraphDot

pub trait GraphDot {
    type NodeType;
    type EdgeType;
    fn configure(&self) -> String;
    fn nodes(&self) -> Vec<Self::NodeType>;
    fn edges(&self) -> Vec<Self::EdgeType>;
    fn get_node(&self, n: usize) -> Option<&Self::NodeType>;
    fn node_attrs(&self, &Self::NodeType) -> DotAttrBlock;
    fn edge_attrs(&self, &Self::EdgeType) -> DotAttrBlock;
    fn edge_source(&self, &Self::EdgeType) -> usize;
    fn edge_target(&self, &Self::EdgeType) -> usize;

    fn node_cluster(&self, _: usize) -> usize { ... }
    fn node_skip(&self, &Self::NodeType) -> bool { ... }
    fn edge_skip(&self, &Self::EdgeType) -> bool { ... }
}

This trait enables graphs to be generated from implementors.

Associated Types

type NodeType

type EdgeType

Required Methods

fn configure(&self) -> String

fn nodes(&self) -> Vec<Self::NodeType>

fn edges(&self) -> Vec<Self::EdgeType>

fn get_node(&self, n: usize) -> Option<&Self::NodeType>

fn node_attrs(&self, &Self::NodeType) -> DotAttrBlock

fn edge_attrs(&self, &Self::EdgeType) -> DotAttrBlock

fn edge_source(&self, &Self::EdgeType) -> usize

fn edge_target(&self, &Self::EdgeType) -> usize

Provided Methods

fn node_cluster(&self, _: usize) -> usize

Nodes with the same node_cluster return value will be put in the same graphviz-cluster.

fn node_skip(&self, &Self::NodeType) -> bool

fn edge_skip(&self, &Self::EdgeType) -> bool