Struct radeco_lib::middle::regfile::SubRegisterFile

pub struct SubRegisterFile {
    pub whole_registers: Vec<ValueType>,
    pub whole_names: Vec<String>,
    // some fields omitted
}

A structure containing information about whole and partial registers of a platform. Upon creation it builds a vector of ValueTypes representing whole registers to be added to a PhiPlacer.

It can then translate accesses to partial registers to accesses of whole registers. Shifts and masks are added automatically.

Fields

whole_registers

ValueTypes of whole registers ready to be added to a PhiPlacer. The index within PhiPlacer to the first register is needed as argument base to read_register and write_register later.

whole_names

Contains the respective names for the registers described in whole_registers

Methods

impl SubRegisterFile

fn new(reg_info: &LRegInfo) -> SubRegisterFile

Creates a new SubRegisterFile based on a provided register profile.

fn write_register<'a, T: SSAMod<BBInfo=BBInfo> + 'a>(&self, phiplacer: &mut PhiPlacer<'a, T>, base: usize, block: T::ActionRef, var: &String, value: T::ValueRef)

Emit code for setting the specified register to the specified value. Will automatically insert code for shifting and masking in case of subregisters. This implies that it also tries to read the old value of the whole register.

Arguments

phiplacer - A PhiPlacer that has already been informed of our variables. It will also give us access to the SSA to modify

base - Index of the PhiPlacer variable that corresponds to our first register.

block - Reference to the basic block to which operations will be appended.

var - Name of the register to write as a string.

value - An SSA node whose value shall be assigned to the register. As with most APIs in radeco, we will not check if the value is reachable from the position where the caller is trying to insert these operations.

fn read_register<'a, T: SSAMod<BBInfo=BBInfo> + 'a>(&self, phiplacer: &mut PhiPlacer<'a, T>, base: usize, block: T::ActionRef, var: &String) -> T::ValueRef

Emit code for reading the current value of the specified register.

Arguments

phiplacer - A PhiPlacer that has already been informed of our variables. It will also give us access to the SSA to modify

base - Index of the PhiPlacer variable that corresponds to our first register.

block - Reference to the basic block to which operations will be appended.

var - Name of the register to read as a string.

Return value

A SSA node representing the value of the register as seen by the current end of block. Unless prior basic blocks are marked as sealed in the PhiPlacer this will always return a reference to a Phi node. Either way, once nodes are sealed redundant Phi nodes are eliminated by PhiPlacer.