Module radeco_lib::middle::ssa::cfg_traits

Defines the traits to be implemented by the Control Flow Graph (CFG).

These traits act as a way to abstract the actual storage mechanism of the CFG (and SSA). Any struct that implement these traits can be used with other methods.

Design

CFG - This acts as the base trait upon which the whole SSA form is build. The CFG trait provides accessors to the Basic Blocks of the program as well as the Control Flow edges that connect these blocks.

CFGMod - This trait provides methods to manipulate the Control Flow Graph.

The above traits are generic over indexes that are used to refer to edges and nodes in a CFG. CFG::ActionRef represents the type that is used to reference the Basic Blocks in the Control Flow Graph. Similarly, CFG::CFEdgeRef represents the type that is used to reference edges in the graph.

It is important to note that the trait SSA requires CFG to be implemented. The underlying CFG trait continues to provide access to the control flow structure of the program even in the SSA form. In essence, the SSA graph can be thought of as a composition (or superposition) of the SSA graph and the CFG graph.

Individual traits and their methods are explained in their respective docs.

Note: Reference in the docs refers to any type that is used to index nodes and edges in the graph and not necessarily pointers.

Traits

CFG
Provides accessors to the underlying storage

CFGMod
Provides mutators to the underlying storage