Trait radeco_lib::middle::ssa::cfg_traits::CFG

pub trait CFG {
    type ActionRef: Eq + Hash + Clone + Copy + Debug;
    type CFEdgeRef: Eq + Hash + Clone + Copy + Debug;
    fn blocks(&self) -> Vec<Self::ActionRef>;
    fn start_node(&self) -> Self::ActionRef;
    fn exit_node(&self) -> Self::ActionRef;
    fn get_unconditional(&self, i: &Self::ActionRef) -> Self::ActionRef;
    fn preds_of(&self, node: Self::ActionRef) -> Vec<Self::ActionRef>;
    fn succs_of(&self, node: Self::ActionRef) -> Vec<Self::ActionRef>;
    fn invalid_action(&self) -> Self::ActionRef;
    fn edges_of(&self, i: &Self::ActionRef) -> Vec<Self::CFEdgeRef>;
    fn incoming_edges(&self, i: &Self::ActionRef) -> Vec<Self::CFEdgeRef>;
    fn info(&self, i: &Self::CFEdgeRef) -> (Self::ActionRef, Self::ActionRef);
    fn find_edge(&self, source: &Self::ActionRef, target: &Self::ActionRef) -> Self::CFEdgeRef;
    fn true_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef;
    fn false_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef;
    fn next_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef;
    fn invalid_edge(&self) -> Self::CFEdgeRef;

    fn source_of(&self, i: &Self::CFEdgeRef) -> Self::ActionRef { ... }
    fn target_of(&self, i: &Self::CFEdgeRef) -> Self::ActionRef { ... }
}

Provides accessors to the underlying storage

Associated Types

type ActionRef: Eq + Hash + Clone + Copy + Debug

type CFEdgeRef: Eq + Hash + Clone + Copy + Debug

Required Methods

fn blocks(&self) -> Vec<Self::ActionRef>

Reference to all blocks in the CFG

fn start_node(&self) -> Self::ActionRef

Reference to entry block of the CFG

fn exit_node(&self) -> Self::ActionRef

Reference to exit block of the CFG

fn get_unconditional(&self, i: &Self::ActionRef) -> Self::ActionRef

Reference to the next block in the natural flow of the CFG

fn preds_of(&self, node: Self::ActionRef) -> Vec<Self::ActionRef>

Reference to immediate predecessors of block

fn succs_of(&self, node: Self::ActionRef) -> Vec<Self::ActionRef>

Reference to immediate successors of block

fn invalid_action(&self) -> Self::ActionRef

Reference that represents and Invalid block

fn edges_of(&self, i: &Self::ActionRef) -> Vec<Self::CFEdgeRef>

Reference to all outgoing edges from a block

fn incoming_edges(&self, i: &Self::ActionRef) -> Vec<Self::CFEdgeRef>

Reference to all the incoming edges to a block

fn info(&self, i: &Self::CFEdgeRef) -> (Self::ActionRef, Self::ActionRef)

Returns (source_block, target_block) for an edge

fn find_edge(&self, source: &Self::ActionRef, target: &Self::ActionRef) -> Self::CFEdgeRef

Reference to the edge that connects the source to the target.

fn true_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef

Reference to the true edge

fn false_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef

Reference to the false edge

fn next_edge_of(&self, i: &Self::ActionRef) -> Self::CFEdgeRef

Reference to the unconditional edge that flows out of the block

fn invalid_edge(&self) -> Self::CFEdgeRef

Reference that represents an Invalid control flow edge.

Provided Methods

fn source_of(&self, i: &Self::CFEdgeRef) -> Self::ActionRef

Reference to the source block for the edge

fn target_of(&self, i: &Self::CFEdgeRef) -> Self::ActionRef

Reference to the target block for the edge