Trait radeco_lib::middle::ssa::ssa_traits::SSA

pub trait SSA: CFG {
    type ValueRef: Eq + Hash + Clone + Copy + Debug;
    fn exprs_in(&self, i: &Self::ActionRef) -> Vec<Self::ValueRef>;
    fn is_expr(&self, i: &Self::ValueRef) -> bool;
    fn get_phis(&self, i: &Self::ActionRef) -> Vec<Self::ValueRef>;
    fn get_uses(&self, i: &Self::ValueRef) -> Vec<Self::ValueRef>;
    fn get_block(&self, i: &Self::ValueRef) -> Self::ActionRef;
    fn get_operands(&self, i: &Self::ValueRef) -> Vec<Self::ValueRef>;
    fn get_sparse_operands(&self, i: &Self::ValueRef) -> Vec<(u8, Self::ValueRef)>;
    fn get_node_data(&self, i: &Self::ValueRef) -> Result<NodeData, Box<Debug>>;
    fn is_selector(&self, i: &Self::ValueRef) -> bool;
    fn selector_of(&self, i: &Self::ActionRef) -> Option<Self::ValueRef>;
    fn selects_for(&self, i: &Self::ValueRef) -> Self::ActionRef;
    fn get_target(&self, i: &Self::ValueRef) -> Self::ActionRef;
    fn get_branches(&self, i: &Self::ValueRef) -> (Self::ActionRef, Self::ActionRef);
    fn get_true_branch(&self, i: &Self::ValueRef) -> Self::ActionRef;
    fn get_false_branch(&self, i: &Self::ValueRef) -> Self::ActionRef;
    fn args_of(&self, node: Self::ValueRef) -> Vec<Self::ValueRef>;
    fn uses_of(&self, node: Self::ValueRef) -> Vec<Self::ValueRef>;
    fn registers_at(&self, i: &Self::ActionRef) -> Self::ValueRef;
    fn refresh(&self, node: Self::ValueRef) -> Self::ValueRef;
    fn invalid_value(&self) -> Self::ValueRef;
    fn to_value(&self, Self::ActionRef) -> Self::ValueRef;
    fn to_action(&self, Self::ValueRef) -> Self::ActionRef;
    fn node_count(&self) -> usize;
    fn edge_count(&self) -> usize;

    fn lhs(&self, i: &Self::ValueRef) -> Self::ValueRef { ... }
    fn rhs(&self, i: &Self::ValueRef) -> Self::ValueRef { ... }
    fn block_of(&self, i: &Self::ValueRef) -> Self::ActionRef { ... }
}

Trait for the SSA Form implementation.

Associated Types

type ValueRef: Eq + Hash + Clone + Copy + Debug

Required Methods

fn exprs_in(&self, i: &Self::ActionRef) -> Vec<Self::ValueRef>

Get all the NodeIndex of all operations/expressions in the BasicBlock with index 'i'.

fn is_expr(&self, i: &Self::ValueRef) -> bool

Check if the node at the given index is a expression or not.

fn get_phis(&self, i: &Self::ActionRef) -> Vec<Self::ValueRef>

Get all phis in the BasicBlock with index 'i'.

fn get_uses(&self, i: &Self::ValueRef) -> Vec<Self::ValueRef>

Get all the uses of the node with index 'i'.

fn get_block(&self, i: &Self::ValueRef) -> Self::ActionRef

Get the NodeIndex of the BasicBlock to which node with index 'i' belongs to.

fn get_operands(&self, i: &Self::ValueRef) -> Vec<Self::ValueRef>

Get the operands for the operation with NodeIndex 'i'.

fn get_sparse_operands(&self, i: &Self::ValueRef) -> Vec<(u8, Self::ValueRef)>

Get the operands for the operation with NodeIndex 'i' as tuples.

fn get_node_data(&self, i: &Self::ValueRef) -> Result<NodeData, Box<Debug>>

Get the actual NodeData.

fn is_selector(&self, i: &Self::ValueRef) -> bool

Returns true if the expression acts as a Selector for control flow.

fn selector_of(&self, i: &Self::ActionRef) -> Option<Self::ValueRef>

Returns the selector for the Block.

fn selects_for(&self, i: &Self::ValueRef) -> Self::ActionRef

Get the Block for which i acts as a selector.

fn get_target(&self, i: &Self::ValueRef) -> Self::ActionRef

Get Jump target of a call or an unconditional jump.

fn get_branches(&self, i: &Self::ValueRef) -> (Self::ActionRef, Self::ActionRef)

Get branches of a selector (false_branch, true_branch).

fn get_true_branch(&self, i: &Self::ValueRef) -> Self::ActionRef

Helper method that gets only the true branch.

fn get_false_branch(&self, i: &Self::ValueRef) -> Self::ActionRef

Helper method that gets only the false branch.

fn args_of(&self, node: Self::ValueRef) -> Vec<Self::ValueRef>

Gets the data dependencies of a value node in any order. (See get_operands for ordered return value)

fn uses_of(&self, node: Self::ValueRef) -> Vec<Self::ValueRef>

Gets uses dependents of a value node. (Equivalent as SSAMod::get_uses)

fn registers_at(&self, i: &Self::ActionRef) -> Self::ValueRef

Get a node that has all register values at the beginning of the specified basic block as args

fn refresh(&self, node: Self::ValueRef) -> Self::ValueRef

Updates a node reference to the latest version in case of replacement

fn invalid_value(&self) -> Self::ValueRef

fn to_value(&self, Self::ActionRef) -> Self::ValueRef

fn to_action(&self, Self::ValueRef) -> Self::ActionRef

fn node_count(&self) -> usize

fn edge_count(&self) -> usize

Provided Methods

fn lhs(&self, i: &Self::ValueRef) -> Self::ValueRef

Get the lhs() of the Operation with NodeIndex 'i'.

fn rhs(&self, i: &Self::ValueRef) -> Self::ValueRef

Get the rhs() of the Operation with NodeIndex 'i'.

fn block_of(&self, i: &Self::ValueRef) -> Self::ActionRef

Get the NodeIndex of the BasicBlock to which node with index 'i' belongs to. (Alias for get_block)