Trait radeco_lib::middle::ssa::ssa_traits::SSAMod

pub trait SSAMod: SSA + CFGMod {
    fn add_op(&mut self, block: Self::ActionRef, opc: MOpcode, vt: ValueType) -> Self::ValueRef;
    fn add_const(&mut self, block: Self::ActionRef, value: u64) -> Self::ValueRef;
    fn add_phi(&mut self, block: Self::ActionRef, vt: ValueType) -> Self::ValueRef;
    fn add_undefined(&mut self, block: Self::ActionRef, vt: ValueType) -> Self::ValueRef;
    fn add_comment(&mut self, block: Self::ActionRef, vt: ValueType, msg: String) -> Self::ValueRef;
    fn mark_selector(&mut self, node: Self::ValueRef, block: Self::ActionRef);
    fn phi_use(&mut self, phi: Self::ValueRef, node: Self::ValueRef);
    fn phi_unuse(&mut self, phi: Self::ValueRef, node: Self::ValueRef);
    fn op_use(&mut self, node: Self::ValueRef, index: u8, argument: Self::ValueRef);
    fn replace(&mut self, node: Self::ValueRef, replacement: Self::ValueRef);
    fn remove(&mut self, node: Self::ValueRef);
    fn cleanup(&mut self);
    fn remove_edge(&mut self, i: &Self::CFEdgeRef);
}

Trait for modifying SSA data

Required Methods

fn add_op(&mut self, block: Self::ActionRef, opc: MOpcode, vt: ValueType) -> Self::ValueRef

Add a new operation node.

fn add_const(&mut self, block: Self::ActionRef, value: u64) -> Self::ValueRef

Add a new constant node.

fn add_phi(&mut self, block: Self::ActionRef, vt: ValueType) -> Self::ValueRef

Add a new phi node.

fn add_undefined(&mut self, block: Self::ActionRef, vt: ValueType) -> Self::ValueRef

Add a new undefined node

fn add_comment(&mut self, block: Self::ActionRef, vt: ValueType, msg: String) -> Self::ValueRef

Add a new comment node

fn mark_selector(&mut self, node: Self::ValueRef, block: Self::ActionRef)

Mark the node as selector for the control edges away from the specified basic block

fn phi_use(&mut self, phi: Self::ValueRef, node: Self::ValueRef)

Add a data source to a phi node.

fn phi_unuse(&mut self, phi: Self::ValueRef, node: Self::ValueRef)

Remove a data source from a phi node.

fn op_use(&mut self, node: Self::ValueRef, index: u8, argument: Self::ValueRef)

Set the index-th argument of the node.

fn replace(&mut self, node: Self::ValueRef, replacement: Self::ValueRef)

Replace one node by another within one basic block.

fn remove(&mut self, node: Self::ValueRef)

Remove a node without replacement

fn cleanup(&mut self)

Perform a cleanup. (Will invalidate indices)

fn remove_edge(&mut self, i: &Self::CFEdgeRef)

Remove conrtol flow edge. This is a part of SSAMod as this potentially modifies the ssa.